requirements.txt
pillow
imagemacgick #brew install imagemagick